Move common helper to element_maps module
